make forge work
